Rename actions/post_utils to actions/new_post and move to TS
