+++ actions/new_post.test.js
@@ -7,7 +7,7 @@ import configureStore from 'redux-mock-store';
 import {receivedNewPost} from 'mattermost-redux/actions/posts';
 import {Posts} from 'mattermost-redux/constants';
 
-import * as PostActionsUtils from 'actions/post_utils';
+import * as NewPostActions from 'actions/new_post';
 import {Constants} from 'utils/constants';
 
 const mockStore = configureStore([thunk]);
@@ -26,7 +26,7 @@ global.Date.now = jest.fn(() => POST_CREATED_TIME);
 const INCREASED_POST_VISIBILITY = {amount: 1, data: 'current_channel_id', type: 'INCREASE_POST_VISIBILITY'};
 const STOP_TYPING = {type: 'stop_typing', data: {id: 'current_channel_idundefined', now: POST_CREATED_TIME, userId: 'some_user_id'}};
 
-describe('actions/post_utils', () => {
+describe('actions/new_post', () => {
     const latestPost = {
         id: 'latest_post_id',
         user_id: 'current_user_id',
@@ -83,7 +83,7 @@ describe('actions/post_utils', () => {
         const newPost = {id: 'new_post_id', channel_id: 'current_channel_id', message: 'new message', type: Constants.PostTypes.ADD_TO_CHANNEL, user_id: 'some_user_id', create_at: POST_CREATED_TIME, props: {addedUserId: 'other_user_id'}};
         const websocketProps = {team_id: 'team_id', mentions: ['current_user_id']};
 
-        await testStore.dispatch(PostActionsUtils.completePostReceive(newPost, websocketProps));
+        await testStore.dispatch(NewPostActions.completePostReceive(newPost, websocketProps));
         expect(testStore.getActions()).toEqual([
             INCREASED_POST_VISIBILITY,
             {
@@ -124,11 +124,11 @@ describe('actions/post_utils', () => {
 
             window.isActive = true;
 
-            await testStore.dispatch(PostActionsUtils.setChannelReadAndViewed(post2, {}));
+            await testStore.dispatch(NewPostActions.setChannelReadAndViewed(post2, {}));
 
             expect(testStore.getActions()).toEqual([{
                 type: 'MOCK_MARK_CHANNEL_AS_READ',
-                args: [channelId, null, true],
+                args: [channelId, undefined, true],
             }, {
                 type: 'MOCK_MARK_CHANNEL_AS_VIEWED',
                 args: [channelId],
@@ -164,7 +164,7 @@ describe('actions/post_utils', () => {
 
             window.isActive = false;
 
-            await testStore.dispatch(PostActionsUtils.setChannelReadAndViewed(post2, {}));
+            await testStore.dispatch(NewPostActions.setChannelReadAndViewed(post2, {}));
 
             expect(testStore.getActions()).toEqual([{
                 type: 'MOCK_MARK_CHANNEL_AS_UNREAD',
@@ -203,7 +203,7 @@ describe('actions/post_utils', () => {
 
             window.isActive = true;
 
-            await testStore.dispatch(PostActionsUtils.setChannelReadAndViewed(post2, {}));
+            await testStore.dispatch(NewPostActions.setChannelReadAndViewed(post2, {}));
 
             expect(testStore.getActions()).toEqual([{
                 type: 'MOCK_MARK_CHANNEL_AS_UNREAD',
@@ -240,11 +240,11 @@ describe('actions/post_utils', () => {
                 },
             });
 
-            await testStore.dispatch(PostActionsUtils.setChannelReadAndViewed(post2, {}));
+            await testStore.dispatch(NewPostActions.setChannelReadAndViewed(post2, {}));
 
             expect(testStore.getActions()).toEqual([{
                 type: 'MOCK_MARK_CHANNEL_AS_READ',
-                args: [channelId, null, false],
+                args: [channelId, undefined, false],
             }, {
                 type: 'MOCK_MARK_CHANNEL_AS_VIEWED',
                 args: [channelId],
@@ -280,7 +280,7 @@ describe('actions/post_utils', () => {
                 },
             });
 
-            await testStore.dispatch(PostActionsUtils.setChannelReadAndViewed(post2, {}));
+            await testStore.dispatch(NewPostActions.setChannelReadAndViewed(post2, {}));
 
             expect(testStore.getActions()).toEqual([{
                 type: 'MOCK_MARK_CHANNEL_AS_UNREAD',
@@ -317,7 +317,7 @@ describe('actions/post_utils', () => {
                 },
             });
 
-            await testStore.dispatch(PostActionsUtils.setChannelReadAndViewed(post2, {}));
+            await testStore.dispatch(NewPostActions.setChannelReadAndViewed(post2, {}));
 
             expect(testStore.getActions()).toEqual([{
                 type: 'MOCK_MARK_CHANNEL_AS_UNREAD',
